+++ zscene/src/sprite.rs
@@ -205,9 +205,7 @@ impl Sprite {
                     data.pos.y(),
                     data.color,
                     DrawTextureParams {
-                        dest_size: Some(
-                            data.scale * Vec2::new(texture.width() as f32, texture.height() as f32),
-                        ),
+                        dest_size: Some(data.scale * Vec2::new(texture.width(), texture.height())),
                         ..Default::default()
                     },
                 );
